adding in ruby versions with prompt, prompt for ibm

+++ ibmcloud/install.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 # Ref: https://console.bluemix.net/docs/cli
 
-echo "Installing IBM Cloud CLI Tools..."
-curl -sL https://ibm.biz/idt-installer | bash
-ibmcloud plugin install container-service -r 'IBM Cloud'
-ibmcloud plugin install cloud-databases
+echo
+read -p "About to install IBM Cloud CLI Tools, are you sure you want this? [y/N]" -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  echo "Installing IBM Cloud CLI Tools..."
+  curl -sL https://ibm.biz/idt-installer | bash
+  ibmcloud plugin install container-service -r 'IBM Cloud'
+  ibmcloud plugin install cloud-databases
+fi
+
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+  echo "Skipping IBM Cloud CLI Tools..."
+fi
